requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2
pyyaml==6.0.2
httpx==0.27.2
plotly==5.23.0
streamlit==1.37.1
pandas==2.2.2
python-dateutil==2.9.0.post0
